df_load

This function load and preprocess a dataset from a CSV file, split the date frame into 2 parts that the first part is nice formatted and the second part need manipulation, correct the wrong row entries in the second part, convert the numeric columns into correct data type, combine the data together.

df_load(
  url,
  skip1,
  skip2,
  n_max1,
  n_max2,
  error_line,
  error_record,
  correct_bef_error_record,
  val_corrected,
  error_col,
  predicted_factor
)

Arguments

url

The quoted URL or file path of the CSV file.

skip1

The number of rows to skip at the beginning of the file for the first splited dataset.

skip2

The number of rows to skip at the beginning of the file for the second splited dataframe.

n_max1

The maximum number of rows to read for the first splited dataframe.

n_max2

The maximum number of rows to read for the second splited dataframe.

error_line

The row index containing errors that need to be corrected.

error_record

A vector of the column indices containing errors that need to be corrected.

correct_bef_error_record

The number of correctly formatted rows before the row containing errors.

val_corrected

list of corrected entries to the error_line.

error_col

A vector of the corrected values for the error columns.

predicted_factor

The quoted column name to be converted to a factor.

Value

The preprocessed data frame, with missing values corrected and data types converted.

Examples

df_load(
  url = "https://raw.githubusercontent.com/mwaskom/seaborn-data/master/iris.csv",
  skip1 = 0,
  skip2 = 0,
  n_max1 = 3,
  n_max2 = 2,
  error_line = 2,
  error_record = 2:4,
  correct_bef_error_record = 1,
  val_corrected = list(3.5, 1.4, 0.2, "setosa"),
  error_col = c("petal_width"),
  predicted_factor = "species"
)
#> # A tibble: 5 × 5
#>   sepal_length sepal_width petal_length petal_width species
#>          <dbl>       <dbl>        <dbl>       <dbl> <fct>  
#> 1          5.1         3.5          1.4         0.2 setosa 
#> 2          4.9         3.5          1.4         0.2 setosa 
#> 3          5.1         3.5          1.4         0.2 setosa 
#> 4          4.9         3            1.4         0.2 setosa 
#> 5          4.7         3.2          1.3         0.2 setosa